cover: handle upgrade banner
