+++ extensions/shared/plan-utils.js
@@ -88,6 +88,11 @@ export function isUpgradable( name ) {
 		return false;
 	}
 
+	// core/cover is handled in ./extensions/shared/blocks/cover
+	if ( name === 'core/cover' ) {
+		return false;
+	}
+
 	let blockName = /^jetpack\//.test( name )
 		? name.substr( 8, name.length )
 		: name;
